requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
Django==5.1.3
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.5.4
django-otp-yubikey==1.1.0
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-two-factor-auth==1.17.0
frozenlist==1.5.0
greenlet==3.1.1
gTTS==2.5.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.0
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.142
multidict==6.1.0
numpy==1.26.4
openai==1.47.0
orjson==3.10.11
packaging==24.2
phonenumbers==8.13.50
pillow==10.4.0
propcache==0.2.0
psycopg==3.2.2
psycopg-binary==3.2.2
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pypng==0.20220715.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==9.0.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
whitenoise==6.7.0
yarl==1.17.1
YubiOTP==1.0.0
